Prepare obj pool and singleton

+++ Assets/00Game/01Script/ICanGetHit.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface ICanGetHit 
 {
-    void GetHit(float damage);
+    void GetHit(int damage);
 }
